migracion
registros mediante pruebas unitarias de empleado , administrado,
acudientes

+++ database/migrations/2017_04_12_073146_create_totalAnio.php
@@ -20,7 +20,7 @@ class CreateTotalAnio extends Migration
             $table->string('anio',45)->nullable();
             $table->dateTime('fechaGeneracion')->nullable();
             $table->integer('idAdministrador');
-            $table->integer('totalAniocol')->nullable();
+            $table->double('totalAniocol',16,8)->nullable();
             $table->timestamps();
         
         });
